+++ controllers/contacts-controller.js
@@ -1,4 +1,4 @@
-const { Contact } = require("../models/contact");
+const { Contact } = require("../models/contact-model");
 const { HttpError, ctrlWrapper } = require("../helpers");
 
 const getAll = async (req, res) => {
