+++ routes/api/contacts-routes.js
@@ -1,8 +1,8 @@
 const express = require("express");
 
-const ctrl = require("../../controllers/contacts");
+const ctrl = require("../../controllers/contacts-controller");
 const { validateBody, isValidId } = require("../../middlewares");
-const schemas = require("../../models/contact");
+const schemas = require("../../models/contact-model");
 
 const router = express.Router();
